pyproject.toml
[project]
name = "tesseract-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Sarah",email = "vanwars@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.30.0,<1.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
